requirements.txt
pytest>=7.0
python-dotenv>=1.0
sentence-transformers>=3.0
openai>=1.0
chromadb>=0.4.0
google-generativeai>=0.8.0
